+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **GM-994** - Add currency data on the store config
 - **GM-995** - Remove dependence useSession
 - **GM-996** - Get locale from the store config on the login, register and changing password flows
+- **GM-997** - Use currency code from the store config on events
 
 ### Fixed